Add a maximum version constraint on `redis`
If this gem is used in conjunction with `redis` 3.2.0 it won't throw a
runtime error exactly, but the internals have changed enough that it
doesn't do what it's supposed to anymore and instead of connecting to
the master host suggested by Sentinel, will connect to the `redis`
default host on `localhost:6379`.

This patch puts a maximum version constraint on `redis` of 3.1.0, which
in my tests is still compatible.

This took one of our services down pretty hard, and hopefully this patch
and a new version of `redis-sentinel` will save someone else from the
same fate.